// models/notification.model.ts
export interface NotificationDto {
  notificationMasterId: number;
  inventoryId: number;
  userId: number;
  action: string;
  category: number;
  isRead: boolean;
  actionType: number;
  updatedDate: string;
  inventory?: Inventory;
  appointment?: Appointment;
}

interface Inventory {
  inventoryId: number;
  purchaseDate: string;
  vin: string;
  color: string;
  locationFrom: string;
  iYear: number;
  vehicleModelId: number;
  vehicleMakelId: number;
  saleDate: null;
  salePrice: null;
  soldTo: string;
  typeOfSale: string;
  originalCost: number;
  isRegistration: boolean;
  status: string;
  notes: string;
  isPaid: string;
  bankId: null;
  title: string;
  purchaseFromId: number;
  soldById: null;
  buyerId: number;
  floorCost: null;
  dealerFee: null;
  lotId: number;
  isOnline: boolean;
  isInspection: boolean;
  announcement: string;
  battery: string;
  checkEngineLight: string;
  ac: string;
  windows: string;
  tireLight: string;
  radio: string;
  sunroof: string;
  windsheild: string;
  upholstrey: string;
  cleanliness: string;
  anyLights: string;
  amythingMike: string;
  others: string;
  keyNo: string;
  numberOfKeys: number;
  salePriceCash: null;
  salePriceWire: null;
  salePriceCheque: null;
  salePriceDescription: string;
  salePriceChequeNo: string;
  isLicensePlate: boolean;
  testDriveId: number;
  photoUrl: string;
  features: string;
  stock: number;
  mileage: number;
  createdById: number;
  createdDate: string;
  updatedById: number;
  updatedDate: string;
  bodyWork: string;
  salePriceZelle: number;
  isPaymentDone: boolean;
  firstPaymentDueDate: null;
  isPaymentReceived: boolean;
  isTitleScanned: boolean;
  isContractScanned: boolean;
  isAllExpensesAdded: boolean;
  onlineCost: number;
  reducedCost: number;
  customerName: string;
  customerPhoneNumber: null;
  titleRegisterationFee: null;
  documentFee: null;
  saleTax: null;
  isWholeSale: boolean;
  isRepairShop: null;
  days: number;
  onlineDays: string;
  make: string;
  model: string;
  purchaseFrom: string;
  totalOriginalCost: number;
  totalAddedCost: number;
  profit: number;
  soldBy: null;
  imageName: null;
}

interface Appointment {
  appointmentId: number;
  inventoryId: number;
  customerName: string;
  customerAddress: string;
  contactNo: string;
  email: string;
  referredBy: string;
  appointmentDate: string;
  notes: string;
  leadId: null;
  createdById: number;
  createdDate: string;
  updatedById: number;
  updatedDate: string;
  action: string;
  tblCustomerImages: any[];
  tblTestDrives: any[];
  vehicleInfo: null;
  sendSMS: null;
  message: null;
}